Fix project

+++ api_yamdb/api/serializers.py
@@ -63,7 +63,7 @@ class TitleSerializer(serializers.ModelSerializer):
     category = SlugToDictField(
         slug_field="slug", queryset=Category.objects.all()
     )
-    rating = serializers.FloatField(
+    rating = serializers.IntegerField(
         source='reviews__score__avg', read_only=True, allow_null=True
     )
 
@@ -81,7 +81,7 @@ class TitleSerializer(serializers.ModelSerializer):
 class TitleGetSerializer(serializers.ModelSerializer):
     genre = GenreSerializer(many=True, read_only=True)
     category = CategorySerializer(read_only=True)
-    rating = serializers.FloatField(
+    rating = serializers.IntegerField(
         source='reviews__score__avg', read_only=True, allow_null=True
     )
